requirements.txt
numpy>=2.0.0
langchain>=0.3.0,<0.4.0
langchain-community>=0.3.0,<0.4.0
langchain-ollama>=0.2.0,<0.3.0
langchain-chroma>=0.1.4,<0.3.0
langchain-core>=0.3.0,<0.4.0
chromadb>=0.5.5
pypdf>=4.3.1
python-dotenv>=1.0.1
streamlit>=1.39.0
